\name{concatenate}
\Rdversion{1.1}
\alias{concatenate}
\title{concatenate}
\description{###Concatenate multi-dimensional arrays along the given dimension, a ###generalization of cbind().  Arrays must have the same dimensions ###except gor the dimension along which the binding occurs. Attempts ###to preserve dimnames; if more than one argument provides names for ###a dimension, the earlier argument takes precedence.  Arrays may ###have different numbers of dimensions in which case they are taken ###to have length 1 in all training dimensions. Vectors without a dim ###attribute are taken to be column vectors.}
\usage{
concatenate(..., along = N, fill = FALSE, match.names = FALSE)
}
\arguments{
  \item{\dots}{
}
  \item{along}{
}
  \item{fill}{
}
  \item{match.names}{
}
}
\details{
}
\value{###The bound array.}
\references{
}
\author{Peter Meilstrup <peterm@shadlen.org>}
\note{
}


\seealso{
}
\examples{}