\name{extend}
\Rdversion{1.1}
\alias{extend}
\title{extend}
\description{###extend (or truncate) a multidimensional array by adding NA entries ###at the end or each dimension (or truncate by deleting.) The number ###of dimensions of a must equal the length of d.}
\usage{
extend(a, d)
}
\arguments{
  \item{a}{
}
  \item{d}{
}
}
\details{
}
\value{
}
\references{
}
\author{Peter Meilstrup <peterm@shadlen.org>}
\note{
}


\seealso{
}
\examples{}