\name{manipulations-package}
\Rdversion{1.1}
\alias{manipulations-package}
\docType{package}
\title{Various Data Manipulations}
\description{A collection of data manupulation functions that are useful to the authors.}
\details{
\tabular{ll}{Package: \tab manipulations\cr
Version: \tab 0.1\cr
Date: \tab 2009-07-29\cr
Title: \tab Various Data Manipulations\cr
Author: \tab Peter Meilstrup <peterm@shadlen.org>\cr
Maintainer: \tab Peter Meilstrup <peterm@shadlen.org>\cr
Depends: \tab R (>= 2.9), RUnit\cr
Description: \tab A collection of data manupulation functions that are useful to the authors.\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://www.r-project.org/\cr}
}
\author{Peter Meilstrup <peterm@shadlen.org>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}